projects/kafka-core/pyproject.toml
[tool.poetry]
name = "kafka-core"
version = "0.0.0"
description = "Kafka core module."
authors = ["Anthony Caliani <github.com/avcaliani>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
